refactor: remove prop-types
