+++ package.json
@@ -51,7 +51,6 @@
     "lodash.uniqueid": "^4.0.1",
     "lodash.merge": "^4.6.2",
     "pretty-bytes": "^5.6.0",
-    "prop-types": "^15.7.2",
     "qr-scanner": "^1.2.0",
     "qrcode.react": "^1.0.1",
     "query-string": "^7.0.0",
@@ -59,7 +58,6 @@
     "react-dom": "^17.0.2",
     "react-dropzone": "^11.3.2",
     "react-feather": "^2.0.9",
-    "react-google-recaptcha": "^2.1.0",
     "react-helmet": "^6.1.0",
     "react-markdown": "^6.0.2",
     "react-redux": "^7.2.4",
@@ -136,7 +134,6 @@
     "puppeteer": "^10.0.0",
     "react-test-renderer": "^17.0.2",
     "redux-devtools-extension": "^2.13.9",
-    "sinon": "^11.1.1",
     "style-loader": "^2.0.0",
     "tailwindcss": "^2.1.4",
     "testcafe": "^1.14.2",
