--- src/components/QrReader/index.js
@@ -1,3 +1,0 @@
-import QrReaderZone from "./qrReader";
-
-export default QrReaderZone;
