+++ src/components/QrReader/qrReader.test.tsx
